Implemented deleteIdleStats and friends
Also implemented/fixed all the delete* config directives,
deleteCounters, ... which at the time I wrote the initial version
of Net::Statsd::Server were compatible with the nodejs version,
but haven't kept us since then.

Now they are fully equivalent. Also added a really nice integration
test about this. Inject a gauge, see it in the flushed data,
and check that it disappears when deleteIdleStats is in effect.